Potential fix for code scanning alert no. 2: Comparison of narrow type with wide type in loop condition

+++ src/test/java/io/github/arthurhoch/kissbinary/rinha/RinhaBinaryValidator.java
@@ -73,7 +73,7 @@ final class RinhaBinaryValidator {
     static boolean readLabel(byte[] data, int vectorIndex, HeaderInfo header) {
         BinaryReader reader = BinaryReader.from(data, RinhaBinaryFormat.ENDIANNESS);
         long labelOffset = RinhaBinaryFormat.labelDataOffset(header.vectorCount);
-        for (int i = 0; i < labelOffset; i++) {
+        for (long i = 0; i < labelOffset; i++) {
             reader.readByte();
         }
         long[] labelWords = reader.readLongArray(header.labelWordCount);
